HerDesign de GUI

+++ src/main/java/be/personal/pieces/piece.java
@@ -1,16 +1,16 @@
 package be.personal.pieces;
 
 public abstract class piece{
-    public byte[] pos;
+    public int[] pos;
     public boolean moved;
     public color co;
 
     protected piece(int x, int y, color c){
         assert x > 0 && x < 9;
         assert y > 0 && y < 9;
-        pos = new byte[2];
-        pos[1] = (byte) x;
-        pos[0] = (byte) y;
+        pos = new int[2];
+        pos[1] = x;
+        pos[0] = y;
         co = c;
         moved = true;
     }
